requirements.txt
asgiref==3.8.1 
django==5.0.7 
sqlparse==0.5.1 
typing-extensions==4.12.2
djangorestframework==3.15.2
django-rest-swagger==2.2.0
drf-spectacular==0.26.5
numpy==2.0.1 
environs==11.0.0 
marshmallow==3.21.3 
packaging==24.1 
python-dotenv==1.0.1